feat: hide mobile menu on desktop
